[Refactor] : service 레이어 생성
Git : #16
