+++ api/eb_fast_api/domain/auth/login/sources/login_schema.py
@@ -10,5 +10,7 @@ class Token(BaseModel):
         accessToken: str,
         refreshToken: str,
     ):
-        self.accessToken = accessToken
-        self.refreshToken = refreshToken
+        super().__init__(
+            accessToken=accessToken,
+            refreshToken=refreshToken,
+        )
