+++ api/eb_fast_api/snippets/tests/dictionary_test.py
@@ -1,4 +1,4 @@
-from eb_fast_api.snippets.dictionary.sources import dictionary
+from eb_fast_api.snippets.sources import dictionary
 
 
 def test_safeDict():
